Automatically create data in DB on startup
This works by means of the docker-entrypoint-initdb.d feature of the
mongo Docker image:

- https://hub.docker.com/_/mongo
- https://github.com/docker-library/docs/blob/master/mongo/README.md#initializing-a-fresh-instance

This feature runs scripts in the /docker-entrypoint-initdb.d directory
when the container is created. These scripts can do things like
initialising the database with default data.

The scripts can be either MongoDB JavaScript files or shell scripts.

JavaScript files (ending in .js) have to be written as explained here:
https://docs.mongodb.com/manual/tutorial/write-scripts-for-the-mongo-shell/.
They are executed as 'mongo <js-file>', as defined in the code here:
https://github.com/docker-library/mongo/blob/master/4.1/docker-entrypoint.sh#L321.

Shell scripts (ending in .sh) are sourced, as defined in the code here:
https://github.com/docker-library/mongo/blob/master/4.1/docker-entrypoint.sh#L320

The problem with using JavaScript files is that they do not allow to
read environment variables (see discussion here https://github.com/docker-library/mongo/issues/257).
However, with shell scripts this is easily possible. For this reason,
the current solution uses a shell script (init-data.sh) to insert the
default data into the database.

The init-data.sh script is included in the database container image by
creating a new custom image on top of the MongoDB image that simply adds
the init-data.sh file to the /docker-entrypoint-initdb.d directory.

The database and collection to use is defined in the config-db.yaml
ConfigMap, which is used to initialise env variables for the database
pod (as well as for the webserver pod).